+++ NEWS.md
@@ -8,6 +8,12 @@ __New Vignettes__
 
 - [Time Series Clustering:](https://business-science.github.io/timetk/articles/TK09_Clustering.html) Uses the new `tk_tsfeatures()` function to perform time series clustering. #95 #84
 
+__Improvements__
+
+Improvements for __point forecasting__ when the target is n-periods into the future.
+
+- `time_series_cv()`, `time_series_split()`: New parameter `point_forecast`. This is useful for testing / assessing the n-th prediction in the future. When set to `TRUE`, will return a single point that returns on the last value in `assess`.  
+
 # timetk 2.6.2
 
 __Improvements__